run symlinks script to simplify rebase
