+++ splink/athena/athena_comparison_library.py
@@ -1,7 +1,12 @@
-from .athena_helpers.athena_comparison_imports import (  # noqa: F401
-    array_intersect_at_sizes,
-    distance_function_at_thresholds,
-    distance_in_km_at_thresholds,
-    exact_match,
-    levenshtein_at_thresholds,
+import warnings
+
+from ..exceptions import SplinkDeprecated
+from .comparison_library import *  # noqa: F403
+
+warnings.warn(
+    "Importing directly from `splink.athena.athena_comparison_library` "
+    "is deprecated and will be removed in Splink v4. "
+    "Please import from `splink.athena.comparison_library` going forward.",
+    SplinkDeprecated,
+    stacklevel=2,
 )
