+++ splink/athena/athena_comparison_template_library.py
@@ -1,5 +1,12 @@
-# Not all Comparison Template Library functions are currently implemented
-# for Athena due to limited string matching capability in
-# cll.comparison_level_library
+import warnings
 
-from .athena_helpers.athena_comparison_imports import postcode_comparison  # noqa: F401
+from ..exceptions import SplinkDeprecated
+from .comparison_template_library import *  # noqa: F403
+
+warnings.warn(
+    "Importing directly from `splink.athena.athena_comparison_template_library` "
+    "is deprecated and will be removed in Splink v4. "
+    "Please import from `splink.athena.comparison_template_library` going forward.",
+    SplinkDeprecated,
+    stacklevel=2,
+)
