+++ splink/athena/comparison_template_library.py
@@ -1,0 +1,5 @@
+# Not all Comparison Template Library functions are currently implemented
+# for Athena due to limited string matching capability in
+# cll.comparison_level_library
+
+from .athena_helpers.athena_comparison_imports import postcode_comparison  # noqa: F401
