+++ splink/duckdb/comparison_template_library.py
@@ -1,0 +1,6 @@
+from .duckdb_helpers.duckdb_comparison_imports import (  # noqa: F401
+    date_comparison,
+    forename_surname_comparison,
+    name_comparison,
+    postcode_comparison,
+)
