+++ splink/duckdb/duckdb_comparison_level_library.py
@@ -1,17 +1,12 @@
-from ..comparison_level_composition import and_, not_, or_  # noqa: F401
-from .duckdb_helpers.duckdb_comparison_imports import (  # noqa: F401
-    array_intersect_level,
-    columns_reversed_level,
-    damerau_levenshtein_level,
-    datediff_level,
-    distance_function_level,
-    distance_in_km_level,
-    else_level,
-    exact_match_level,
-    jaccard_level,
-    jaro_level,
-    jaro_winkler_level,
-    levenshtein_level,
-    null_level,
-    percentage_difference_level,
+import warnings
+
+from ..exceptions import SplinkDeprecated
+from .comparison_level_library import *  # noqa: F403
+
+warnings.warn(
+    "Importing directly from `splink.duckdb.duckdb_comparison_level_library` "
+    "is deprecated and will be removed in Splink v4. "
+    "Please import from `splink.duckdb.comparison_level_library` going forward.",
+    SplinkDeprecated,
+    stacklevel=2,
 )
