+++ splink/spark/comparison_library.py
@@ -1,0 +1,12 @@
+from .spark_helpers.spark_comparison_imports import (  # noqa: F401
+    array_intersect_at_sizes,
+    damerau_levenshtein_at_thresholds,
+    datediff_at_thresholds,
+    distance_function_at_thresholds,
+    distance_in_km_at_thresholds,
+    exact_match,
+    jaccard_at_thresholds,
+    jaro_at_thresholds,
+    jaro_winkler_at_thresholds,
+    levenshtein_at_thresholds,
+)
