+++ splink/spark/comparison_template_library.py
@@ -1,0 +1,6 @@
+from .spark_helpers.spark_comparison_imports import (  # noqa: F401
+    date_comparison,
+    forename_surname_comparison,
+    name_comparison,
+    postcode_comparison,
+)
