+++ splink/spark/spark_comparison_library.py
@@ -1,12 +1,12 @@
-from .spark_helpers.spark_comparison_imports import (  # noqa: F401
-    array_intersect_at_sizes,
-    damerau_levenshtein_at_thresholds,
-    datediff_at_thresholds,
-    distance_function_at_thresholds,
-    distance_in_km_at_thresholds,
-    exact_match,
-    jaccard_at_thresholds,
-    jaro_at_thresholds,
-    jaro_winkler_at_thresholds,
-    levenshtein_at_thresholds,
+import warnings
+
+from ..exceptions import SplinkDeprecated
+from .comparison_library import *  # noqa: F403
+
+warnings.warn(
+    "Importing directly from `splink.spark.spark_comparison_library` "
+    "is deprecated and will be removed in Splink v4. "
+    "Please import from `splink.spark.comparison_library` going forward.",
+    SplinkDeprecated,
+    stacklevel=2,
 )
