+++ splink/sqlite/comparison_library.py
@@ -1,0 +1,5 @@
+from .sqlite_helpers.sqlite_comparison_imports import (  # noqa: F401
+    distance_function_at_thresholds,
+    exact_match,
+    levenshtein_at_thresholds,
+)
