+++ splink/sqlite/sqlite_comparison_level_library.py
@@ -1,10 +1,12 @@
-from ..comparison_level_composition import and_, not_, or_  # noqa: F401
-from .sqlite_helpers.sqlite_comparison_imports import (  # noqa: F401
-    columns_reversed_level,
-    distance_function_level,
-    else_level,
-    exact_match_level,
-    levenshtein_level,
-    null_level,
-    percentage_difference_level,
+import warnings
+
+from ..exceptions import SplinkDeprecated
+from .comparison_level_library import *  # noqa: F403
+
+warnings.warn(
+    "Importing directly from `splink.sqlite.sqlite_comparison_level_library` "
+    "is deprecated and will be removed in Splink v4. "
+    "Please import from `splink.sqlite.comparison_level_library` going forward.",
+    SplinkDeprecated,
+    stacklevel=2,
 )
